Split up auto configuration cleanly
