+++ logbook-spring-boot-starter/src/main/java/org/zalando/logbook/spring/LogbookAutoConfiguration.java
@@ -1,6 +1,7 @@
 package org.zalando.logbook.spring;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.apache.http.client.HttpClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,12 +12,14 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
 import org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration;
+import org.springframework.boot.autoconfigure.security.SecurityFilterAutoConfiguration;
 import org.springframework.boot.autoconfigure.web.WebMvcAutoConfiguration;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.Ordered;
+import org.springframework.security.web.SecurityFilterChain;
 import org.zalando.logbook.BodyFilter;
 import org.zalando.logbook.BodyFilters;
 import org.zalando.logbook.ChunkingHttpLogWriter;
@@ -39,7 +42,10 @@ import org.zalando.logbook.RawResponseFilter;
 import org.zalando.logbook.RawResponseFilters;
 import org.zalando.logbook.RequestFilter;
 import org.zalando.logbook.ResponseFilter;
+import org.zalando.logbook.httpclient.LogbookHttpRequestInterceptor;
+import org.zalando.logbook.httpclient.LogbookHttpResponseInterceptor;
 import org.zalando.logbook.servlet.LogbookFilter;
+import org.zalando.logbook.servlet.Strategy;
 
 import javax.servlet.Filter;
 import java.util.List;
@@ -59,26 +65,11 @@ import static javax.servlet.DispatcherType.REQUEST;
 })
 public class LogbookAutoConfiguration {
 
-    private static final String AUTHORIZED = "authorizedLogbookFilter";
-
     private final LogbookProperties properties;
 
     @Autowired
     public LogbookAutoConfiguration(final LogbookProperties properties) {
         this.properties = properties;
-    }
-
-    @Bean
-    @ConditionalOnWebApplication
-    @ConditionalOnProperty(name = "logbook.filter.enabled", havingValue = "true", matchIfMissing = true)
-    @ConditionalOnMissingBean(name = AUTHORIZED)
-    public FilterRegistrationBean authorizedLogbookFilter(final Logbook logbook) {
-        final Filter filter = new LogbookFilter(logbook);
-        final FilterRegistrationBean registration = new FilterRegistrationBean(filter);
-        registration.setName(AUTHORIZED);
-        registration.setDispatcherTypes(REQUEST, ASYNC, ERROR);
-        registration.setOrder(Ordered.LOWEST_PRECEDENCE);
-        return registration;
     }
 
     @Bean
@@ -208,5 +199,71 @@ public class LogbookAutoConfiguration {
     public Logger httpLogger() {
         return LoggerFactory.getLogger(properties.getWrite().getCategory());
     }
+
+    @Configuration
+    @ConditionalOnClass({
+            HttpClient.class,
+            LogbookHttpRequestInterceptor.class,
+            LogbookHttpResponseInterceptor.class
+    })
+    static class HttpClientAutoConfiguration {
+
+        @Bean
+        @ConditionalOnMissingBean(LogbookHttpRequestInterceptor.class)
+        public LogbookHttpRequestInterceptor logbookHttpRequestInterceptor(final Logbook logbook) {
+            return new LogbookHttpRequestInterceptor(logbook);
+        }
+
+        @Bean
+        @ConditionalOnMissingBean(LogbookHttpResponseInterceptor.class)
+        public LogbookHttpResponseInterceptor logbookHttpResponseInterceptor() {
+            return new LogbookHttpResponseInterceptor();
+        }
+
+    }
+
+    @Configuration
+    @ConditionalOnClass(Filter.class)
+    @ConditionalOnWebApplication
+    static class ServletFilterConfiguration {
+
+        private static final String FILTER_NAME = "authorizedLogbookFilter";
+
+        @Bean
+        @ConditionalOnProperty(name = "logbook.filter.enabled", havingValue = "true", matchIfMissing = true)
+        @ConditionalOnMissingBean(name = FILTER_NAME)
+        public FilterRegistrationBean authorizedLogbookFilter(final Logbook logbook) {
+            final Filter filter = new LogbookFilter(logbook);
+            final FilterRegistrationBean registration = new FilterRegistrationBean(filter);
+            registration.setName(FILTER_NAME);
+            registration.setDispatcherTypes(REQUEST, ASYNC, ERROR);
+            registration.setOrder(Ordered.LOWEST_PRECEDENCE);
+            return registration;
+        }
+
+    }
+
+    @Configuration
+    @ConditionalOnClass(SecurityFilterChain.class)
+    @ConditionalOnWebApplication
+    @AutoConfigureAfter(SecurityFilterAutoConfiguration.class)
+    static class SecurityServletFilterConfiguration {
+
+        private static final String FILTER_NAME = "unauthorizedLogbookFilter";
+
+        @Bean
+        @ConditionalOnProperty(name = "logbook.filter.enabled", havingValue = "true", matchIfMissing = true)
+        @ConditionalOnMissingBean(name = FILTER_NAME)
+        public FilterRegistrationBean unauthorizedLogbookFilter(final Logbook logbook) {
+            final Filter filter = new LogbookFilter(logbook, Strategy.SECURITY);
+            final FilterRegistrationBean registration = new FilterRegistrationBean(filter);
+            registration.setName(FILTER_NAME);
+            registration.setDispatcherTypes(REQUEST, ASYNC, ERROR);
+            registration.setOrder(Ordered.HIGHEST_PRECEDENCE + 1);
+            return registration;
+        }
+
+    }
+
 
 }
