+++ logbook-spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,1 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.zalando.logbook.spring.LogbookAutoConfiguration,\
-org.zalando.logbook.spring.LogbookSecurityAutoConfiguration,\
-org.zalando.logbook.spring.LogbookHttpClientAutoConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration = org.zalando.logbook.spring.LogbookAutoConfiguration
