+++ logbook-spring-boot-starter/src/test/java/org/zalando/logbook/spring/Application.java
@@ -5,8 +5,6 @@ import org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration;
 
 @ImportAutoConfiguration({
         LogbookAutoConfiguration.class,
-        LogbookSecurityAutoConfiguration.class,
-        LogbookHttpClientAutoConfiguration.class,
         JacksonAutoConfiguration.class
 })
 public class Application {
